Resolved bug where hitting the delete key would bring up the prompt to delete root nodes and putty nodes

+++ mRemoteV1/UI/Window/ConnectionTreeWindow.cs
@@ -299,6 +299,7 @@ namespace mRemoteNG.UI.Window
 
         public void DeleteSelectedNode()
         {
+            if (SelectedNode is RootNodeInfo || SelectedNode is PuttySessionInfo) return;
             if (!UserConfirmsDeletion()) return;
             ConnectionTreeModel.DeleteNode(SelectedNode);
             Runtime.SaveConnectionsBG();
